requirements.txt
Jinja2~=3.1.2
PyYAML~=6.0
fastapi~=0.116.1
python-jose[cryptography]~=3.5.0
uvicorn[standard]~=0.35.0
passlib~=1.7.4
bcrypt~=4.3.0
pydantic~=2.11.7
python-multipart~=0.0.6
passlib~=1.7.4
requests~=2.32.4
urllib3~=2.5.0
paramiko>=3.3.1
httpx>=0.25.1
